move encoding to preprocessing
