--- docs/encode/ehrapy.api.encode.encode.rst
@@ -1,12 +1,0 @@
-﻿:github_url: ehrapy.api.encode.encode
-
-ehrapy.api.encode.encode
-========================
-
-.. autofunction:: ehrapy.api.encode.encode
-
-.. _sphx_glr_backref_ehrapy.api.encode.encode:
-
-.. minigallery:: ehrapy.api.encode.encode
-   :add-heading: Gallery
-   :heading-level: -
