--- docs/encode/ehrapy.api.encode.type_overview.rst
@@ -1,12 +1,0 @@
-﻿:github_url: ehrapy.api.encode.type_overview
-
-ehrapy.api.encode.type\_overview
-================================
-
-.. autofunction:: ehrapy.api.encode.type_overview
-
-.. _sphx_glr_backref_ehrapy.api.encode.type_overview:
-
-.. minigallery:: ehrapy.api.encode.type_overview
-   :add-heading: Gallery
-   :heading-level: -
