--- docs/encode/ehrapy.api.encode.undo_encoding.rst
@@ -1,12 +1,0 @@
-﻿:github_url: ehrapy.api.encode.undo_encoding
-
-ehrapy.api.encode.undo\_encoding
-================================
-
-.. autofunction:: ehrapy.api.encode.undo_encoding
-
-.. _sphx_glr_backref_ehrapy.api.encode.undo_encoding:
-
-.. minigallery:: ehrapy.api.encode.undo_encoding
-   :add-heading: Gallery
-   :heading-level: -
