+++ ehrapy/__init__.py
@@ -12,4 +12,4 @@ ehrapy_pypi_latest.check_latest()
 
 traceback.install(width=200, word_wrap=True)
 
-from ehrapy.api import data, encode, io, plot, preprocessing, tools
+from ehrapy.api import data, io, plot, preprocessing, tools
