+++ ehrapy/api/__init__.py
@@ -4,7 +4,7 @@ from ehrapy.api._util import print_versions
 settings: EhrapyConfig = ehrapy_settings
 
 from ehrapy.api import data as dt
-from ehrapy.api import encode as ec
+from ehrapy.api import io
 from ehrapy.api import plot as pl
 from ehrapy.api import preprocessing as pp
 from ehrapy.api import tools as tl
