+++ ehrapy/api/preprocessing/encoding/_categoricals.py
